refactor(marshallers): changes the type ids in marshallers of "Pokemon", "Move", "Species" and "Type"

+++ src/main/java/com/pokemon/ao/persistence/marshaller/SpeciesMarshaller.java
@@ -30,7 +30,7 @@ public class SpeciesMarshaller implements Marshaller<SpeciesVO, Species>{
 
     @Override
     public SpeciesVO unmarshall(Species species) {
-        Long id = species.getId();
+        Integer id = species.getId();
         String name = species.getName();
         String sprite = species.getSpriteUrl();
         TypeVO type = typeMarshaller.unmarshall(species.getType());
